refactor: 채팅방 멤버 명령 책임 분리 (#609)
* refactor: 채팅방 멤버 명령 책임 분리

- ChatService에 남아 있던 방 나가기와 멤버 강퇴 명령 흐름을 ChatRoomMemberCommandService로 이동
- direct 나가기는 leftAt 상태 변경으로 유지하고 group 나가기는 멤버 삭제 정책을 그대로 둠
- club group 나가기 거부, 비그룹 강퇴 거부, 자기 자신과 방장 강퇴 거부 정책을 기존 응답 코드와 함께 보존

* refactor: 채팅방 멤버 조회 정책 중복 제거

- ChatService와 ChatRoomMemberCommandService가 같은 멤버 조회 실패 정책을 재사용하도록 공용 helper로 분리
- Repository 기본 메서드는 Mockito 단위 테스트 stubbing과 맞지 않아 피하고, 기존 조회 쿼리 호출 경로를 유지
- 접근 정책/에러코드 변경 시 두 서비스가 서로 다른 예외를 던지는 회귀를 방지

* refactor: 채팅방 멤버 조회 helper 이름 정리

- 공용 조회 helper 이름을 서비스 호출부의 도메인 용어와 맞춰 읽기 쉽게 조정
- 조회 실패 시 FORBIDDEN_CHAT_ROOM_ACCESS로 변환하는 기존 정책은 그대로 유지
- 기능 변경 없이 리뷰 대응 코드의 의도를 더 직접적으로 드러냄
